Complete 10_fibonacci
I wasn't clever enough to come up with the simple algorithm for this
exercise, i had to be handed it. Maths is not my strong suit.
Not too disappointed about it though since once provided with the
algorith it was simple to pass all the tests. I'm here to learn how to
code not how to maths.