Complete 12_findTheOldest
This one was pretty tough, spent a few hours on it. In the end i was
able to come up with a decent solution then condense it down with my
first uses of ternary operators. I honestly much prefer my solution to
the provided one. Don't know if it's worse or not though.